Move test types to test_types package
