+++ openapi/generator_test.go
@@ -3,6 +3,7 @@ package openapi
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/wI2L/fizz/testdata/test_types"
 	"io/ioutil"
 	"math"
 	"reflect"
@@ -25,56 +26,6 @@ var genConfig = &SpecGenConfig{
 }
 
 var rt = reflect.TypeOf
-
-type (
-	W struct {
-		A, B string
-	}
-	u struct {
-		S int
-	}
-	q int
-	X struct {
-		*X // ignored, recursive embedding
-		*Y
-		A string `validate:"required"`
-		B *int
-		C bool `deprecated:"true"`
-		D []*Y
-		E [3]*X
-		F *X
-		G *Y
-		H map[int]*Y // ignored, unsupported keys type
-		*u
-		uu *u // ignored, unexported field
-		q     // ignored, embedded field of non-struct type
-		*Q
-		*V `json:"data"`
-	}
-	Y struct {
-		H float32   `validate:"required"`
-		I time.Time `format:"date"`
-		J *uint8    `deprecated:"oui"` // invalid value, interpreted as false
-		K *Z        `validate:"required"`
-		N struct {
-			Na, Nb string
-			Nc     time.Duration
-		}
-		l int // ignored
-		M int `json:"-"`
-	}
-	Z map[string]*Y
-	Q struct {
-		NnNnnN string `json:"nnNnnN"`
-	}
-	V struct {
-		L int
-	}
-)
-
-func (*X) TypeName() string { return "XXX" }
-func (*W) Format() string   { return "wallet" }
-func (*W) Type() string     { return "string" }
 
 // TestStructFieldName tests that the name of a
 // struct field can be correctly extracted.
@@ -190,7 +141,7 @@ func TestSchemaFromComplex(t *testing.T) {
 	g := gen(t)
 	g.UseFullSchemaNames(false)
 
-	sor := g.newSchemaFromType(rt(new(X)))
+	sor := g.newSchemaFromType(rt(new(test_types.X)))
 	assert.NotNil(t, sor)
 
 	b, err := json.Marshal(sor)
@@ -465,7 +416,7 @@ func TestAddOperation(t *testing.T) {
 			},
 		},
 	}
-	_, err := g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(Z{}), infos)
+	_, err := g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(test_types.Z{}), infos)
 	if err != nil {
 		t.Error(err)
 	}
@@ -508,7 +459,7 @@ func TestAddOperation(t *testing.T) {
 	}
 	// Try to add the operation again with the same
 	// identifier. Expected to fail.
-	_, err = g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(Z{}), infos)
+	_, err = g.AddOperation(path, "POST", "Test", reflect.TypeOf(&In{}), reflect.TypeOf(test_types.Z{}), infos)
 	assert.NotNil(t, err)
 
 	// Add an operation with a bad input type.
@@ -524,25 +475,25 @@ func TestTypeName(t *testing.T) {
 		t.Error(err)
 	}
 	// Typer interface.
-	name := g.typeName(rt(new(X)))
+	name := g.typeName(rt(new(test_types.X)))
 	assert.Equal(t, "XXX", name)
 
 	// Override. This has precedence
 	// over the interface implementation.
-	err = g.OverrideTypeName(rt(new(X)), "")
+	err = g.OverrideTypeName(rt(new(test_types.X)), "")
 	assert.NotNil(t, err)
-	assert.Equal(t, "XXX", g.typeName(rt(new(X))))
+	assert.Equal(t, "XXX", g.typeName(rt(new(test_types.X))))
 
-	g.OverrideTypeName(rt(new(X)), "xXx")
-	assert.Equal(t, "xXx", g.typeName(rt(X{})))
+	g.OverrideTypeName(rt(new(test_types.X)), "xXx")
+	assert.Equal(t, "xXx", g.typeName(rt(test_types.X{})))
 
-	err = g.OverrideTypeName(rt(new(X)), "YYY")
+	err = g.OverrideTypeName(rt(new(test_types.X)), "YYY")
 	assert.NotNil(t, err)
 
 	// Default.
-	assert.Equal(t, "OpenapiY", g.typeName(rt(new(Y))))
+	assert.Equal(t, "Test_typesY", g.typeName(rt(new(test_types.Y))))
 	g.UseFullSchemaNames(false)
-	assert.Equal(t, "Y", g.typeName(rt(Y{})))
+	assert.Equal(t, "Y", g.typeName(rt(test_types.Y{})))
 
 	// Unnamed type.
 	assert.Equal(t, "", g.typeName(rt(struct{}{})))
@@ -708,18 +659,18 @@ func TestOverrideSchema(t *testing.T) {
 	g := gen(t)
 
 	// Type is mandatory.
-	err := g.OverrideDataType(rt(W{}), "", "wallet")
+	err := g.OverrideDataType(rt(test_types.W{}), "", "wallet")
 	assert.NotNil(t, err)
 
 	// Success.
-	err = g.OverrideDataType(rt(&W{}), "string", "wallet")
+	err = g.OverrideDataType(rt(&test_types.W{}), "string", "wallet")
 	assert.Nil(t, err)
 
 	// Data type already overidden.
-	err = g.OverrideDataType(rt(&W{}), "string", "wallet")
+	err = g.OverrideDataType(rt(&test_types.W{}), "string", "wallet")
 	assert.NotNil(t, err)
 
-	sor := g.newSchemaFromType(rt(W{}))
+	sor := g.newSchemaFromType(rt(test_types.W{}))
 	assert.NotNil(t, sor)
 
 	schema := g.resolveSchema(sor)
